pyproject.toml
[tool.poetry]
name = "december-api-tests"
version = "0.1.0"
description = ""
authors = ["alexanderlozovoy <berpress@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pytest = "^6.2.5"
requests = "^2.26.0"
Faker = "^10.0.0"
cattrs = "^1.9.0"
attrs = "^21.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
